+++ ssdb-1.9.2/src/ssdb/ssdb_impl.h
@@ -95,8 +95,8 @@ public:
 	virtual int incr(const Bytes &key, int64_t by, int64_t *new_val);
 	virtual int multi_set(const std::vector<Bytes> &kvs, int offset=0);
 	virtual int multi_del(const std::vector<Bytes> &keys, int offset=0);
-	virtual int setbit(const Bytes &key, int bitoffset, int on);
-	virtual int getbit(const Bytes &key, int bitoffset);
+	virtual int setbit(const Bytes &key, int64_t bitoffset, int on);
+	virtual int getbit(const Bytes &key, int64_t bitoffset);
 	
 	virtual int get(const Bytes &key, std::string *val);
 	virtual int getset(const Bytes &key, std::string *val, const Bytes &newval);
